Extract templates imports
